[step1] 1차 코드리뷰 반영  - README.md 파일 수정  - 패키지 이름 카멜케이스로 수정  - ktlint 설정  - Splitter 인스턴스 메소드 확장함수로 수정  - Splitter object 클래스로 수정  - 테스트 코드 when & then 수정  - 테스트 이름 수정 (커스텀 구분자)
