+++ README.md
@@ -19,7 +19,7 @@
 
 - [x] 쉼표(,) 또는 콜론(:)을 구분자로 문자열을 분리하는 기능
 - [x] 숫자의 합을 출력하는 기능
-- [ ] 커스텀 구분자를 지정하는 기능
+- [x] 커스텀 구분자를 지정하는 기능
 - [x] 빈 문자열이 입력되면 0을 반환하는 기능
 - [x] 숫자나 음수가 입력됐을 때 RuntimeException 을 던지는 기능
 - [x] 덧셈의 결과가 출력되는 기능
