--- src/main/kotlin/AdditionCalculator/view/InputView.kt
@@ -1,8 +1,0 @@
-package AdditionCalculator.view
-
-object InputView {
-  fun getExpression(message: String): String? {
-    println(message)
-    return readLine()
-  }
-}
