--- src/main/kotlin/AdditionCalculator/view/ResultView.kt
@@ -1,5 +1,0 @@
-package AdditionCalculator.view
-
-object ResultView {
-  fun showResult(result: Int) = println(result)
-}
