+++ src/main/kotlin/additionCalculator/view/InputView.kt
@@ -1,0 +1,8 @@
+package additionCalculator.view
+
+object InputView {
+    fun getExpression(message: String): String? {
+        println(message)
+        return readLine()
+    }
+}
