+++ src/main/kotlin/additionCalculator/view/ResultView.kt
@@ -1,0 +1,5 @@
+package additionCalculator.view
+
+object ResultView {
+    fun showResult(result: Int) = println(result)
+}
